backend/requirements.txt
google-generativeai
flask
datetime
requests
flask-cors